refactor patches; fix easyocr download models
